VCL-58 Modified xCAT.pm::does_image_exist. It had been returning undefined if the tmpl file exists but the image files do not, causing new.pm to fail the reservation unnecessarily.
If the tmpl file exists but the image files do not, it now deletes the orphaned tmpl file and returns 0.  If the tmpl file doesn't exist but the image files do, it now creates a tmpl file for the image and returns 1.

Added _create_template and _delete_template subroutines. Changed all locations in xCAT.pm to use _create_template where it had previously been copying the tmpl file inline for consistancy.

Added argument to not display output in ls SSH call in xCAT.pm::retrieve_image..  This command lists remote image library contents. It was adding a lot of useless lines to the log file.

Improved some other comments and log file messages.