VCL-94 Added set_provisioner() and provisioner() subroutines to OS.pm, allowing OS modules to access the provisioning module object.
Added set_os() and os() subroutines to Provisioning.pm, allowing provisioning modules to access the OS module object.

Added code to State.pm::initialize() calling the above-mentioned subroutines. This allows provisioning and OS module objects to access each other via $self->os and $self->provisioner.

Added a block to the beginning of image.pm::process() to call the provisioning module's capture() subroutine if it has been implemented. If it hasn't, the block is ignored and process() behaves exactly as it did before.

Made 2 minor changes in utils.pm. Added run_command to the EXPORT list. Added a check in run_scp_command() for the existence of 'no such file' in the output. This wasn't being caught.