VCL-106
Seperated the critical and information section of notify routine. The critical notices caputure the last 100 lines of the vcld.log file the informational notices are for folks that have emails disabled. By doing this it can shave 20 or more secs off the wait time for the connect button to show.