VCL-112 Added DataStructure.pm::retrieve_user_data(). This subroutine retrieves user data from the database. It adds flexibility in that the DataStructure class can be used to retrieve user data whether or not a reservation exists for that user. It will be used in the near future to retrieve the data about an image owner so that the image affiliation can be determined. Image owner detailed data is not being retrieved when a DataStructure object is created based on a reservation.
There are no hard-coded column names in the subroutine code. The column names are retrieved from the information_schema database.  

The subroutine can be called as either an object method or class function. A default value had to be added to request_data in order for this to work without warning messages. Arguments can either be a user ID or login ID. If no argument is given, the subroutine must be called as an object method and a user ID must already be stored in the DataStructure.