VCL-137 Updated vmware.pm to include subroutines which are part of the provisioning module API specification: -power_status() -power_on() -power_off() -power_reset()
Added block of code to capture() which calls a modular OS module's pre_capture() subroutine instead of performing OS-specific tasks within capture().

Changed a regex in capture to include the possibility of an underscore in an OS name. Capture was failing because the regex didn't properly handle an image name beginning with vmwarewinxp_mod.

Fixed a bug in a regex in get_image_size(). It wasn't getting all of the vmdk file names.

Updated header and footer of vmware.pm to match other files, as well as removing NCSU references.

Added get_vmx() sub which returns various paths and filenames. It's functionality would have had to be done in each of the new power_ subs.