VCL-118 Added delay after a failed attempt to utils.pm:run_ssh_command() and run_scp_command(). Also added ability to pass a hash reference to run_ssh_command(). The list of arguments it accepts is somewhat lengthy and this will make it easier to keep the arguments straight when calling it. It will also make it easier to add additional arguments later on.
Added 'max_attempts' argument to run_ssh_command allowing caller to specify how many times the command will be attempted if it fails. The default is 3. This argument allows caller to control the number of attempts. For example, only 1 attempt may be desired if the caller only wants to know if SSH is responding.

Changed call in Windows_mod.pm::wait_for_ssh() to call run_ssh_command using a hash reference argument. It specifies max_attempts to be 1.