VCL-23 Added subroutines and added calls in pre_capture() to further automate image creation tasks: disable_windows_defender() disable_automatic_updates() disable_security_center_notifications()
Added condition to delete_file() to check for "Circular directory" warning when rm is used. It now catches this and doesn't display a warning. Other deletion methods are then attempted.

Fixed bug in filesystem_entry_exists() if path specified included backslashes. Added regex to convert them to forward slashes before checking if exists. Changed command to enclose path in double quotes instead of single.

Added ipconfig_renew() subroutine and added call to it from enable_dhcp(). This is necessary because the default gateways are sometimes lost when DHCP is enabled.

Added to the POD documentation for a few of the subroutines.