VCL-165 Added check to utils.pm::write_currentimage_txt to make sure $ssh_output is defined before trying to access it. This was causing processes to die and be left in the pending state because the following error occurred: "Can't use an undefined value as an ARRAY reference"
Added "return 0" to the capture() sub in vmware.pm, xCAT.pm, and xCAT21.pm if write_currentimage_txt failed. It wasn't catching this. Not having the correct currentimage.txt saved in an image may cause unnecessary reloads.

Added several checks in vmware.pm to make sure result is defined after running run_ssh_command before attempting to use it. If ssh command failed, the process would die because of a "Can't use an undefined value as an ARRAY reference" error and the reservation would be left in the pending state.