VCL-260 Added Windows.pm::set_file_owner() subroutine. Added calls to set root as the owner of /root in pre_capture() and post_load().
VCL-259
Added call to disable the ntsyslog service in Windows.pm::pre_capture(). This service was used to send event log messages to syslog on the management node and isn't used anymore. The service is no longer being properly configured and can cause problems if it attempts to send messages to an address it can't get to.

VCL-261
Reworked output/exit status checking in Windows.pm filesystem_entry_exists() so that it first checks if the output contains "file not found" rather than checking the exit status first. This should be more reliable.

VCL-269
Reworked Windows.pm::logoff_users() to check for disconnected sessions in addition to logged in sessions.

VCL-262
Updated Windows.pm::apply_security_templates() to use the "/overwrite /quiet" secedit.exe switches instead of "/verbose". This causes the .inf file being applied to overwrite existing settings rather than append to them. It is needed to remove the "log on as a service" permission from overwriting existing accounts with this permission.

Other:
Increased the amount of time Windows.pm::reboot() will wait for ping to respond from 4 to 6 minutes. 4 is too short under some circumstances.

Added check to Windows.pm::copy_capture_configuration_files() to check if the source configuration directory in .../tools on the management node exists before attempting to copy it to the node.

Added path to Windows.pm::clean_hard_drive(): /home/root/%USERPROFILE%. This directory is occasionally created by some Cygwin SSH operation. I'm not sure where it's coming from but the directory is not needed.