VCL-279 Added get_module_info sub to utils.pm. It returns the info from the module table. It is used to retrieve a list of all modules in the database so they can each be checked for the implementation of a setup subroutine.
Added get_current_file_name, get_current_package_name, and get_current_subroutine_name subs to utils.pm. These are mainly for debugging purposes and return information about the code where they are called from.

Reworked the argument parsing section of utils.pm. All arguments are now parsed in utils.pm rather than having some parsed in vcld. Moved help sub from vcld to utils.pm so that it can be called from utils.pm if the -h switch is used or if an argument problem is detected.

Added $DAEMON_MODE and $SETUP_MODE variables to utils.pm. $DAEMON_MODE gets set if -d is specified and $SETUP_MODE gets set if -s is specified. This -d switch used to be contained within vcld to control if daemon mode was to be used.

Changed BEGIN block to INIT in utils.pm. This allows subs defined later in utils.pm (help) to be called within the INIT block.

Removed old, unused $TESTING variable from utils.pm and command line argument.

Added code to notify sub to save and restore the STDOUT and STDERR handles when the sub is entered and left. This allows other parts of the code to write to STDOUT and have it appear on the screen, rather than having all or none of the output written to the log file. This is needed to implement the setup framework, where the user is asked questions. notify messages still get written to the log file, but STDOUT gets written to the screen.

Other:
Temporarily added back disablesshd subroutine. This is still being called from reclaim.pm for lab computers. The logic in reclaim.pm needs to be reworked and then disablesshd can be removed again from utils.pm.