VCL-279 Removed command line argument parsing from vcld. This now resides in utils.pm. Added checks for $SETUP_MODE and $DAEMON_MODE variables which are set in utils.pm.
Moved call to rename_vcld_process to earlier in the code, before daemonize is called and commented out line in daemonize which renamed the process. daemonize was unnecessarily renaming the process because it later got renamed by rename_vcld_process.

Added setup_management_node sub to vcld. This gets a list of the modules in the module table, checks each one for the implementation of a setup sub, and calls setup as needed.