VCL-140 Updated code in DataStructure.pm that handles the variable table. Separated out serialize/deserialize functions from get_variable and set_variable to their own subroutines - yaml_serialize and yaml_deserialize. Modified get_variable and set_variable to make them being called as an object method optional.
VCL-145
Added get_image_affiliation_id sub to DataStructure.pm. This returns the affiliation ID of an image owner. It will be used to determine the product key to use when Windows product keys are stored in the database.

Added get_affiliation_info sub to utils.pm. It returns the contents of the affiliation table.