VCL-145 Created the following new subroutines in Version_5.pm to support the dynamic creation of the sysprep.inf file for Windows XP and Windows Server 2003: run_sysprep() get_sysprep_inf_contents() get_sysprep_inf_mass_storage_section() get_driver_inf_paths() set_device_path_key()
These subroutines utilize the winProductKey table to dynamically inject the correct product key (affiliation-specific or global) into the sysprep.inf file. They also scan the driver files to construct the SysprepMassStorage section of sysprep.inf and to set the DevicePath registry key. As a result, the spdrvscn.exe utility is no longer needed and the template sysprep.inf file does not need to be stored on the management node or in the repository.