VCL-301 Added tools directories for Windows 7.
Added Unattend.xml Sysprep configuration file for Windows 7.

Added Security_Center_Reg_Perms.inf security template which opens the permissions on the registry key which needs to be set to disable Security Center notifications.  Added My_Computer_Reg_Perms.inf security template which opens the permissions on the registry key which needs to be set to customize the name of the "My Computer" icon on the desktop.  Renamed windows_security.inf to match the other policy template files.