VCL-301 Added disable_hibernation() and disable_ceip() (customer experience improvement program) subroutines to Windows.pm and added calls from pre_capture().
Modified Windows.pm::reg_add() to allow 0 to be passed as the registry data argument.

Added DisableSR=1 registry key to Windows.pm::disable_system_restore().  It had only been setting the DisableConfig key.

Updated the SSH firewall subroutines in Version_6.pm.  A single SSH command is executed which consists of multiple netsh.exe commands chained together.  The existing SSH rules need to be deleted and then the desired rule is added in the same SSH command.  It would occasionally hang because the SSH connection is broken when the rules are deleted.  I added the get_firewall_state() sub.  When the SSH firewall rules are modified, the firewall state is first checked -- on or off.  If it's on, it gets turned off, the SSH rules are modified, then turned back on.  This should prevent the SSH process from hanging.  Also updated other firewall subroutines for consistency.