VCL-139
help.php:
-modified printHelpForm to only fill in the name of the user if name fields are defined and to handle magic_quotes being on
-modified submitHelpform to handle magic_quotes being on and to allow name to be empty

requests.php: removed adminEditRequest, confirmAdminEditRequest, and submitAdminEditRequest - these have been commented out for a long time and were referencing user's names. If the functionality was ever needed again, it would probably be implemented very differently; so, it was safe to dump these.

userpreferences.php: modified userpreferences() to only show the Personal section if apprioprate fields (names, email, local affiliation password) are set for the user

privileges.php: modified userLookup to only print name and email info if defined for the user

utils.php: modified main to only print the Hello message if names are defined for the user