VCL-250
After correcting this issue, I discovered that passing an affiliation not in the affiliation table would cause a query error.  testGeneralAffiliation in authentication.php had to be modified to correct this, resulting in getUserInfo and getAffilidAndLogin in utils.php needing to be modified, and anything that calls either of those functions.

shibauth.php: modified updateShibUser - check for return status of getAffilidAndLogin, if -1, return NULL

utils.php:
-modified checkAccess - uncommented check to see if $authtype is empty after trying to match the user's affiliation
-modified validateUserid - return 0 if return status of getAffilidAndLogin is -1
-modified getAffilidAndLogin - pass along return status of test function if it is non-zero
-modified addOwnedResourceGroups - return without modifying anything if getUserInfo returns NULL
-modified getUserInfo - return NULL if getAffilidAndLogin returns -1

privileges.php: modified checkUserHasPriv - return 0 if getUserInfo returns -1

authentication.php: modified testGeneralAffiliation - if getAffiliationID returns NULL, return -1