VCL-298 Added utils::create_management_node_directory().  This provides a consistent way of creating directories.
Updated utils.pm::get_image_info() and get_computer_info() to include additional data that get_request_info() already retrieves.

Added vmprofile.imageid to the data returned from utils.pm::get_vmhost_info().

Other:
Updated format_data to display object type names.  Added "use Scalar::Util 'blessed'" to accomplish this.

Added utils.pm::format_number().  It adds commas and rounds a number to decimal places.

Added open(STDOUT... to utils.pm::notify() so that the log file is guaranteed to be opened when writing to it.  A problem was occurring if the log was rotated without restarting vcld.  After being rotated, the log file remained empty.

Updated regex in utils.pm::getdynamicaddress() to check for a space or '.' after the node name when parsing /etc/hosts.  It was incorrectly finding node names with characters after the name it was supposed to be looking for.  For example: vcla1-10 may be used if looking for vcla1-1.

Updated utils.pm::run_ssh_command() and run_scp_command() to not add the string 'none' to the output if no output was returned.  This string may throw off other subroutines which parse the output and expect something exact.  Rearranged error checking in run_scp_command() so that multiple attempts aren't made if certain errors occur which indicate the connection was made but something is wrong with the SCP arguments.