VCL-298 Updated OS.pm::is_ssh_responding() to display if port 22 or 24 is open.
Added insertloadlog call to wait_for_response() which creates an 'osrespond' log entry.  This line is currently commented out until the schema is updated.  This will be used to indicate when the OS has responded in the reservation status information.

Added update_ssh_known_hosts() subroutine.  This checks the known_hosts file, removes old entries for the node, and adds the correct entry.  This code was duplicated in every provisioning module.  It will eventually be called during OS post_load.

Added update_public_ip_address() subroutine.  This checks if static or dynamic addresses are used, and either sets the static address or updates the dynamic address in the database.  This code was also duplicated in every provisioning module and this will eventually be called during OS post_load().