VCL-363 Removed the code from Linux.pm::changepasswd so that it does not recreate /etc/shadow. Removed utils.pm::changelinuxpassword because it is no longer being called.
VCL-49
Removed the commented-out section from utils.pm::reservation_being_processed that called pgrep to check if a process was already running for the reservation. Replaced this with a call to utils.pm::is_management_node_process_running. Updated is_management_node_process_running. It was not parsing the pgrep arguments correctly.

VCL-298
Added code to VMware.pm to check the image.project value and add additional network interfaces if the project is not 'vcl'. It gets a list of all the networks available on the VM host and adds an adapter for any that match up with the image project name. Added Linux.pm::activate_interfaces which finds and activates network interfaces which don't have a corresponding ifcfg-eth* file. The file is created and the interface is brought up allowing interfaces to be added dynamically during load. This is called by Linux.pm::post_load.

Fixed a bug in VMware.pm::get_active_vmx_file_path. It wasn't working correcly when run during an image capture when a new image was created with a different ID.

VCL-100
Added get_network_configuration, get_private_mac_address, get_public_mac_address, and get_public_ip_address to Linux.pm. Added get_private_interface_name and get_public_interface_name to OS.pm. These subroutines make the Linux code more inline with the Windows code and will allow the utils.pm::getdynamicaddress to be removed.

Fixed a bug in Linux.pm::execute. It was using the computer host name rather than the node name. This caused SSH commands to fail because it was attempting to connect to a full DNS host name configured in the computer table which didn't resolve.