VCL-289 Created Provisioning.pm::retrieve_image subroutine based on the code from the existing xCAT and VMware modules.  Added Semaphore.pm which contains code to create an exclusively locked file and Module.pm::get_semaphore for easy access/creation.  The retrieve_image subroutine creates a Semaphore object before retrieving an image from another management node.  This ensures that only 1 retrieval for a given image happens at a time.  Added get_image_repository_path and get_image_repository_search_paths subroutines to VMware.pm allowing it to use the new retrieve_image subroutine.  Also removed get_lockfile and release_lockfile from VMware.pm. They were put there during development.
VCL-100
Updated Linux.pm::get_file_size to return the bytes actually used rather than the apparent size of a file.  This allows an accurate size to be returned when checking the size of thin vmdk files whose different size values may differ widely.

VCL-298
Updated VMware.pm:
Modified is_vm_persistent to return true if the reservation end time is more than 24 hours in the future.
Updated in get_image_size so that it works correctly if an image name is passed as an argument. This is done by image.pm.
Fixed bug in get_vm_disk_adapter_type.  It was returning IDE if using ESX in some situations which doesn't work.

Other
Made minor change to OS.pm::is_ssh_responding so that it doesn't rely on the exit status of the SSH command.
In utils.pm::run_ssh_command, added '2>&1' so it is applied to both the remote command being run and the local SSH command.  This should allow STDOUT and STDERR output to be captured and synchronized.