VCL-298 VMware.pm: Added code to node_status to check if the OS module's post_load tasks have run if the computer is responding and return 'POST_LOAD' if the computer is up but the tasks have not been run.
Changed how vmx directory is named if the VM is persistent to match the nonpersistent name. This allows a nonpersistent reservation a preloaded VM if that VM was configured for persistent mode.

Added post_maintenance_action subroutine.  This allows VM's to be removed from a host via the website.

vSphere_SDK.pm:
Renamed several subs which are private to the module to begin with an underscore to separate them from other subs.

Added check in copy_virtual_disk if the adapter type argument isn't specified, the source vmdk is configured to use IDE, and the VM host is using ESX. This is overridden to use lsiLogic.

Added code to delete_file to check if the path specified is the root of a datastore.  A critical warning is generated if this occurs and nothing is deleted.

VIM_SSH.pm:
Added code to detect the name of the vim-cmd executable installed on the VM host.  VMware Server 2.x uses vmware-vim-cmd whereas ESXi uses vim-cmd.

Fixed bug in vm_unregister().  It was passing the vmx path rather than the VM ID to vim-cmd.