VCL-145 Updated Windows.pm and Version_5.pm so that Windows 5.x image capture processes don't fail if Sysprep is enabled and no stroage driver .inf files are found.
VCL-298
Updated VMware.pm to remove the VM from the VM host when the VM is unassigned via the website.

Other
Added calls to kill the logon screen saver process in the Windows code before a reboot or shutdown is attempted. This occasionally causes a reboot or shutdown to hang.