VCL-196 Removed the explicit System32 path in all the places cmd.exe is called.  Calling cmd.exe with the full path causes problems if the Windows OS is 64-bit.  In this case, System32 is C:\Windows\Sysnative, pointing to the 64-bit copy of the System32 directory.  When cmd.exe is called in this manner, you're running the native 64-bit cmd.exe shell, which does not have access to the Sysnative link.  As a result, if you pass a Sysnative path to a subroutine which calls cmd.exe, this path doesn't exist with the cmd.exe shell and problems occur.
Updated Windows.pm::create_directory().  Removed unused option which allowed an array of directories to be passed.  This was done to simplify the subroutine.

VCL-390
Added code to Windows.pm::get_public_ip_address to detect if an auto-generated 169.254 IP address is retrieved.  It this is the case and DHCP is being used, the subroutine attempts to renew the public IP address to obtain a valid one.

VCL-391
Added check to see if tsshutdn.exe exists on the computer where shutdown.exe is called.  shutdown.exe often fails on Windows 2003 when called remotely and tsshutdn.exe works.  tsshutdn.exe is called if it exists.  Otherwise, shutdown.exe is called.

Other
Improved output in Windows.pm::delete_file if an error occurs.

Changed return code of Windows.pm::delete_files_by_pattern from undefined to 1 if the base directory does not exist (meaning there's nothing to delete).  This was causing warnings to be displayed during capture when attempting to delete some files.

Added Windows.pm::create_text_file.  It isn't being called by anything yet.