VCL-390 Updated Windows.pm::get_network_configuration to call utils.pm::is_public_ip_address().  This code was duplicated in the Windows.pm subroutine and was missing some additional logic included in utils.pm.  Previously, the Windows subroutine considered 169.254 addresses public and would return the first one it found if the public information was requested.  This caused the wrong information to be returned if another interface with a valid public address exists.
VCL-391
Updated places where tsshutdn.exe or shutdown.exe is called to only call tsshutdn.exe if Windows Server 2003 is used.  I've seen this hang on Windows XP.

Other
Added Windows.pm::clear_event_log.  This is called before a computer is shutdown if Sysprep is not used.  This causes only the events that happen during and after an image is loaded to appear in the Event Viewer, making it easier to determine the events related to that image being loaded and also removing older security events which the reservation user shouldn't see.