VCL-223 Updated Windows.pm::search_and_replace_in_files() to escape special characters in the search pattern argument.  It was not working correctly if the pattern contained special characters.  A line containing "-bash: <error message>" would be displayed.  This wasn't being caught when the output was parsed.  Updated regex to catch this.
Updated Windows.pm::sanitize_file() to not return immediately if it fails to search/replace any of the paths.  It now attempts to sanitize all of the paths and keeps track of any errors encountered.

VCL-402
Added Windows.pm::disable_login_screensaver() and a call in Windows.pm::pre_capture() to call this subroutine if the computer is a VM.