VCL-394 There were several duplicated subroutines in vSphere_SDK.pm and VIM_SSH.pm which parsed file paths.  These were moved to VMware.pm.  The only subroutine needed in the helper modules is _get_datastore_info().  vSphere_SDK.pm and VIM_SSH.pm were updated so this subroutine was named the same and returned a hash reference structured the same.
Changed the way images are saved in the repository if the VM profile disk type is 'localdisk'.  It now converts the image to 2gbsparse when saving to the repository.  Saving the vmdk's using the thin format does not work well because thin becomes thick when copied via SCP and under many other situations.  The vmdk's are converted to thin if necessary when copied from the repository to the VM host.

Added code to allow the image repository path to be mounted on the VM host.  This allows VMware's utilities to be used to copy to/from the repository.  If mounted on the VM host, the vmdk's can be converted while they are copied.

Updated VMware.pm::node_status to work if it isn't called as an object method.  It can now be called normally and passed a computer ID.  If called in this manner, a DataStructure object is created for the computer argument and a VMware provisioning object is created.  The code then calls subroutines in the same manner as if called as an object method.