VCL-422 Updated Windows.pm::clean_hard_drive to detect if cleanmgr.exe exists.  This utility is not installed by default on some newer versions of Windows.  It gets installed with the Desktop Optimization feature.  If not installed, a warning would be displayed during capture.  This update should prevent the warning from appearing.
Also updated to code to launch cleanmgr.exe and immediately return. This should prevent a situation where an image capture process would hang indefinitely if cleanmgr.exe encounters a problem such as a corrupt Recycle Bin.  When this happens, the utility displays a warning on the screen.

Added Windows.pm::is_process_running.  This is called by clean_hard_drive to detect when cleanmgr.exe is finished.  It waits 2 minutes and displays a warning if cleanmgr.exe is still running, then proceeds with the capture.