VCL-424 Updated vSphere.pm:
Moved check to determine if the virtualDiskManager object is available in is_restricted to earlier in the subroutine to save a little time.

Added check in initialize to parse the output of the connection attempt to determine if the attempt failed due to an incorrect username or password.  A warning is displayed if this is the case, a normal debug message otherwise.  Some connection attempts are normal if the vSphere SDK isn't available and SSH is used instead.