VCL-394 Updated subroutines which copy and rename vmdk files to detect if the operation failed because the destination is out of space.  If this occurs a critical notification is sent.  This makes the cause of the problem more obvious.
Updated VMware.pm::capture() to revert the changes it made to the vmdk file name/path if the image fails to be copied to the image repository.  Also added calls to power the VM back on if a capture fails.  This is done to return the state of the VM being captured back to the original before the capture was attempted, making it easier to attempt the capture again.

Added check when vmkfstools is called to detect if the utility prompted for the username.