VCL-394 DataStructure.pm. Removed references to vmprofile.nasshare.  This column was previously removed from the schema.
VMware.pm:
Changed default guestOS value from winvista to windows7 for 32-bit Windows 7.  This was a typo.

Enabled disk space checking in the load() subroutine.  This had previously been added but was commented out during development.

Added line in capture() to disable Sysprep for all image captures.

Updated code which removes existing VMs before a VM is loaded.  Updated remove_existing_vms subroutine.  It was performing redundant checks.  Added _get_vmx_file_path_computer_name subroutine.  This is used to determine the computer name from a directory name.  This code was duplicated in a few places so it was moved into its own subroutine for consistency.

Added commented-out code to prepare_vmx which enables VMware's built-in VNC connectivity.  This will be tested and possibly uncommented in the future.

Fixed bug in capture() if the repository is not mounted on the VM host and SCP has to be used to copy a captured image to the repository.  The copied files were not named correctly.

Updated get_vmx_info to retrieve the VCL computer ID from the vmx file and add it to the returned hash.


Other
Fixed a few log output typos in utils.pm.

Updated DataStructure.pm::get_computer_private_ip_address to attempt to return the privateIPaddress value from the database if it could not be retrieved from /etc/hosts.

Added StrictHostKeyChecking=no option to run_ssh_command.  This should prevent problems if it is not configured in the management node's ssh_config file.

Updated get_computer_ids() to use a regex instead of 2 'LIKE' expressions.  This should improve the handling of computer names that begin the same such as vm-1 and vm-10.

Updated DataStructure.pm::