VCL-435 Created ESXi.pm OS module.  It currently only implements a post_load subroutine in order to allow ESXi loads to work.  Some functions in the Linux.pm post_load subroutine fail under ESXi.
Updated xCAT2.pm to handle ESXi 4.1 scripted installs:
Loosened regex which attempts to locate the "Serving xcat" line in the /var/log/messages file.  It was looking for initrd.img.  The "Serving xcat" line appears for ESXi installs but the file is not named initrd.img.

Updated node_status to return 'RELOAD' if the nodetype table doesn't match currentimage.txt.

Updated regex's in _get_image_template_path to remove periods and the 'i' in ESXi.  The OS source patch for ESXi is 'esxi4.1' but the template files reside in an 'esx' directory.

Added code to create currentimage.txt in Linux.pm::post_load if the OS install type is Kickstart.  This code exists in new.pm but should really only exist in Linux.pm since it's specific to Kickstart installs.

Other
Updated Linux.pm to call wait_for_power_off instead of simply waiting 60 seconds.  This matches the Windows code.