minor changes
modified AJtoggleBlockTime - now we actually delete entries from the request table if the start time has not been reached instead of updating the state to be deleted.  It's been a while since I did this, but I think the reason was that the deleted stated ended up causing the machines to get picked up by the reload algorthim.

modified AJupdateBlockStatus - updated to check for computer being in 'reload' state in addition to 'reloading' state