VCL-350 Added snapshot support to the VMware code. Added VMware.pm::snapshot() subroutine and create_snapshot subroutines in the helper modules.
Changed default virtual disk mode to persistent instead of independent-persistent or independent-nonpersistent. A snapshot is created after the VM is registered but before it is powered on. This causes changes to be written to a delta file in the directory where the vmx resides. This eliminates the need to create a full copy of the vmdk for imaging and long-term reservations. Imaging and long-term reservations now use this mode. A full copy of the vmdk is only created for server requests. Added DataStructure.pm::is_server_request subroutine to tell if a request is a server request or not.

Changed all things named nonpersistent to shared, and persistent to dedicated to avoid confusion.

Other
Added some missing block request values to DataStructure.pm.