VCL-470 Added code contributed by Aaron Coburn included in the Jira issue to add a vSphere_SDK.pm::get_total_space subroutine.
VCL-471
Updated VMware.pm::_get_parent_directory_normal_path() to allow non-datastore paths to be specified. This solves a problem where a path residing on the management node is passed, but the subroutine fails.

VCL-450
Fixed a problem where vSphere_SDK.pm::copy_virtual_disk failed because the vmdk adapter type is not listed in the typical info if a snapshot has been created.

Updated vSphere.pm to cache most of the SDK object types it uses including host, vm, datastore. This prevents the code from having to fetch a new object every time, greatly increasing speed. This also reduced a lot of duplicated code.

Fixed problem in VMware.pm::node_status. It was not accounting for VMs loaded before the snapshot functionality was added which were using nonpersistent mode.