VCL-348 Setting privileges in web gui -- Slightly broken
The 2 items mentioned in the JIRA issue must have been fixed in a previous release.  However, if no node had been previously selected, the default one was not being highlighted.  I corrected that and changed the color of the font for selected nodes to be black instead of white to make it easier to read.