VCL-545 Updated Linux.pm::get_file_size to be able to accept an array of files.
Updated Linux.pm::find_files to allow an argument to be passed specifying the type of search to perform. This allows a regex file pattern to be passed.

Added get_copy_speed_info_string to utils.pm. This displays various file copy performance statistics. It was copied out of VMware.pm and added to utils.pm since it can be used elsewhere. 

Added get_random_mac_address to utils.pm.

Updated utils.pm::format_data to detect an XML string. If XML, a hash is constructed using XML::Simple::XMLin and then displayed.


Other
Updated section of post_load in Linux.pm which attempts to run the vcl_post_load script to check if the script exists on the computer being loaded before trying to execute it.  This is done to prevent unnecessary warning messages from appearing in the vcld.log file.