VCL-450 Commented out MemTrimRate=0, sched.mem.pshare.enable=FALSE, and mainMem.useNamedFile=FALSE values in the code that generates the vmx file. These parameters may have adverse performance affects under ESXi 4.x/5.x.
Fixed logic bug in prepare_vmdk. If a dedicated copy of the vmdk needs to be created, it was only checking if the file exists in the repository and not in the datastore. Updated to check both locations.

Updated get_vm_ethernet_adapter_type to attempt to retrieve the type from the reference vmx file if it exists instead of only using the default adapter type for the OS being loaded.

Added long timeout arguments to the vmkfstools and vmware-vdiskmanager commands which may take a long time.