VCL-450 Added logic to the VMware code to configure the VM CPU(s) according to VM host and guest OS limitations. It now checks the VMware license vCPU limit and won't assign more CPUs to a VM than the host is licensed for. Added get_license_info subroutines to vSphere_SDK.pm and VIM_SSH.pm. It also checks the number of cores the host has and won't assign more than this number. Some guest OS's have CPU socket limitations but can support multicore processors. The code now detects this and won't assign more sockets than the guest allows, but will add multiple cores to each CPU.
Added check to the end of VMware::load to make sure the guest OS has recognized the correct number of CPUs. Some OS's require a reboot in order to detect additional CPUs if the image was captured on a VM with fewer CPUs than the VM being loaded has. If the guest OS reports fewer CPUs than should be detected, the VM is rebooted.

Added snapshot.redoNotWithParent vmx parameter because the location of snapshot delta files changed with ESXi 5.0. They are stored by default with the parent vmdk rather than the VM's working directory where the vmx file resides. This parameter causes the delta file to be located in the vmx directory.

Added feature that allows VMs to use a dedicated datastore if a datastore exists on the VM host named after the VM.